Habit Categories

+++ src/pages/Dashboard/screens/Schedule/components/ScheduleSection/components/Habits/index.jsx
@@ -1,6 +1,6 @@
-import React from 'react'
+import React, {useState} from 'react'
 import MoreMenu from '../../../../../../components/MoreMenu'
-import {ArrowDown} from 'react-feather'
+import {ArrowDown, ChevronDown, ChevronUp} from 'react-feather'
 import { NavLink } from 'react-router-dom'
 import AddButton from '../../../AddButton'
 
@@ -56,23 +56,75 @@ const Habits = () => {
         setAllRoutes({...allRoutes, habit: id})
     }
 
+    const [dropDownDay, setDropDownDay] = useState({day: 'All', open: false})
+
     return (
         <div>
             <div className={styles.slotSection} style={{height: 'calc(100vh - 160px)'}}>
-                {habits.length!==0 ? habits.map((item)=>{
-                    return (
-                        <NavLink onClick={()=>setRoute(item.id)} key={item.id} to={`/schedule/habits/${item.id}`} className={styles.sideSectionSlot} activeClassName={styles.activeSectionSlot} data-title={item.name}>
-                           <div className={styles.slotContent}>
-                               <div style={{backgroundColor: colors[item.color]}}>
-                                    {iconsSvg[item.icon]}
-                                </div>
-                                <p>{item.name}</p>
-                            </div>
-                            <MoreMenu items={[{name: "edit", function: ()=>openEditHabitModal(item)}, {name: "delete", function: ()=>deleteHabit(item.id)}]} id={`scheduleSlotsMoreMenu${item.id}`} pos={{right: '-5vh', top: '3.5vh'}} />
-                            <CheckBtn times={item.times} id={item.id} timesCompleted={item.timesCompleted} datesCompleted={item.datesCompleted} />
-                        </NavLink>
-                    )
-                }) : <div className={styles.helperTextAddEntry}><p>Add your first entry!</p><ArrowDown /></div>}
+                {habits.length===0 ? <div className={styles.helperTextAddEntry}><p>Add your first entry!</p><ArrowDown /></div> : 
+                <div>
+                    <div className={styles.category}>
+                    <div className={styles.dropDown}>
+                        <h3 onClick={()=>setDropDownDay({...dropDownDay, open: !dropDownDay.open})}><span>{dropDownDay.day}</span><ChevronDown /></h3>
+                        {dropDownDay.open?
+                        <ul>
+                            <li onClick={()=>setDropDownDay({day: 'All', open: false})}>All</li>
+                            <li onClick={()=>setDropDownDay({day: 'Monday', open: false})}>Monday</li>
+                            <li onClick={()=>setDropDownDay({day: 'Tuesday', open: false})}>Tuesday</li>
+                            <li onClick={()=>setDropDownDay({day: 'Wednesday', open: false})}>Wednesday</li>
+                            <li onClick={()=>setDropDownDay({day: 'Thursday', open: false})}>Thursday</li>
+                            <li onClick={()=>setDropDownDay({day: 'Friday', open: false})}>Friday</li>
+                            <li onClick={()=>setDropDownDay({day: 'Saturday', open: false})}>Saturday</li>
+                            <li onClick={()=>setDropDownDay({day: 'Sunday', open: false})}>Sunday</li>
+                        </ul>
+                        :null}
+                    </div>
+                    {habits.length!==0 ? habits.map((item)=>{
+                        if(item.times !== item.timesCompleted){
+                            for(let key in item.repeat){
+                                if(item.repeat[key] !== null){
+                                    if(dropDownDay.day.substring(0, 3).toLowerCase() === key){
+                                        return (
+                                            <NavLink onClick={()=>setRoute(item.id)} key={item.id} to={`/schedule/habits/${item.id}`} className={styles.sideSectionSlot} activeClassName={styles.activeSectionSlot} data-title={item.name}>
+                                                <div className={styles.slotContent}>
+                                                <div style={{backgroundColor: colors[item.color]}}>
+                                                        {iconsSvg[item.icon]}
+                                                    </div>
+                                                    <p>{item.name}</p>
+                                                </div>
+                                                <MoreMenu items={[{name: "edit", function: ()=>openEditHabitModal(item)}, {name: "delete", function: ()=>deleteHabit(item.id)}]} id={`scheduleSlotsMoreMenu${item.id}`} pos={{right: '-5vh', top: '3.5vh'}} />
+                                                <CheckBtn times={item.times} id={item.id} timesCompleted={item.timesCompleted} datesCompleted={item.datesCompleted} />
+                                            </NavLink>   
+                                        )
+                                    }
+                                }
+                            }
+                        }
+                        return null
+                    }) : null}
+                    </div>
+                    <div className={styles.category}>
+                        <h3><span>Completed</span><ChevronUp /></h3>
+                        {habits.length!==0 ? habits.map((item)=>{
+                        if(item.times === item.timesCompleted){
+                            return (
+                                <NavLink onClick={()=>setRoute(item.id)} key={item.id} to={`/schedule/habits/${item.id}`} className={styles.sideSectionSlot} activeClassName={styles.activeSectionSlot} data-title={item.name}>
+                                    <div className={styles.slotContent}>
+                                    <div style={{backgroundColor: colors[item.color]}}>
+                                            {iconsSvg[item.icon]}
+                                        </div>
+                                        <p>{item.name}</p>
+                                    </div>
+                                    <MoreMenu items={[{name: "edit", function: ()=>openEditHabitModal(item)}, {name: "delete", function: ()=>deleteHabit(item.id)}]} id={`scheduleSlotsMoreMenu${item.id}`} pos={{right: '-5vh', top: '3.5vh'}} />
+                                    <CheckBtn times={item.times} id={item.id} timesCompleted={item.timesCompleted} datesCompleted={item.datesCompleted} />
+                                </NavLink>   
+                            )
+                        }
+                        return null
+                    }) : null}
+                    </div>
+                </div>
+                }
             </div>
             <AddButton name="habit" onclick={openHabitModal} />
         </div>
